% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuAlloc.R
\name{gpuAlloc}
\alias{gpuAlloc}
\title{A gpuAlloc function}
\usage{
gpuAlloc(device = 1)
}
\arguments{
\item{device}{if you have multi GPUs, choose one by device index.}
}
\description{
this fubction prevent GPU memory shutdown error, when training model on Tensorflow.
}
\keyword{GPU}
\keyword{allocation}
\keyword{growth}
\keyword{memory}
